refactor(facet): extract the run-recording path into facet-record
New MIT crate crates/facet-record: `record(&RecordRequest)` builds the
Lattice row from a resolved request and its HTTP outcome (redaction,
request hash, body placement, machine-store index), plus ConfigOverrides,
open_store, and the FACET_ACTOR / FACET_SESSION / FACET_NO_RECORD helpers.
facet-cli's `request run` now calls it; behavior and JSON unchanged (16
integration tests green on apiary). Recording tests moved with the code.
NOTICE and docs/FACET.md list the crate. Also rustfmt on the CLI tests.
