+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/core",
     "crates/desktop",
     "crates/facet",
+    "crates/facet-record",
     "crates/http",
     "crates/lattice",
     "crates/opencollection",
@@ -18,6 +19,7 @@ default-members = [
     "crates/cli",
     "crates/core",
     "crates/facet",
+    "crates/facet-record",
     "crates/http",
     "crates/lattice",
     "crates/opencollection",
