+++ crates/facet-record/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "facet-record"
+version.workspace = true
+edition.workspace = true
+license = "MIT"
+rust-version.workspace = true
+description = "Facet: shared run-recording path into Lattice for the CLI and TUI"
+
+[dependencies]
+lattice = { path = "../lattice" }
+probe-core = { path = "../core" }
+probe-http = { path = "../http" }
+serde_json = "1"
+
+[lints]
+workspace = true
